requirements.txt
python-dotenv==0.14.0
requests==2.24.0
SQLAlchemy==1.3.18
